requirements.txt
google-genai>=1.0.0
reportlab>=4.0.0
python-dotenv>=1.0.0
rich>=13.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.9
